tooltip attributes for prop group row
