--- packages/client/src/pages/MainPage/containers/AttributesEditor/TooltipBooleanRow/TooltipBooleanRow.tsx
@@ -1,27 +1,0 @@
-import { AttributeIcon } from "components";
-import React from "react";
-import { StyledIconWrap, StyledRow } from "./TooltipBooleanRowStyles";
-
-interface TooltipBooleanRow {
-  attributeName: string;
-  label: string;
-  show: boolean;
-}
-export const TooltipBooleanRow: React.FC<TooltipBooleanRow> = ({
-  attributeName,
-  label,
-  show,
-}) => {
-  return (
-    <>
-      {show && (
-        <StyledRow>
-          <StyledIconWrap>
-            <AttributeIcon attributeName={attributeName} />
-          </StyledIconWrap>
-          {label}
-        </StyledRow>
-      )}
-    </>
-  );
-};
