--- packages/client/src/pages/MainPage/containers/AttributesEditor/TooltipBooleanRow/TooltipBooleanRowStyles.tsx
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const StyledRow = styled.div`
-  display: flex;
-  align-items: center;
-`;
-export const StyledIconWrap = styled.span`
-  margin-right: ${({ theme }) => theme.space[2]};
-`;
